requirements.txt
python-dotenv>=1.0.0
requests>=2.25
Flask==2.3.3
flask-cors==4.0.0
